+++ sample/src/App.tsx
@@ -28,7 +28,7 @@ Sentry.init({
   dsn: SENTRY_INTERNAL_DSN,
   debug: true,
   beforeSend: (e, hint) => {
-    console.log('Event beforeSend:', e, 'hint:', hint);
+    // console.log('Event beforeSend:', e, 'hint:', hint);
     return e;
   },
   // This will be called with a boolean `didCallNativeInit` when the native SDK has been contacted.